Added damage powerup. Health is now as a float. Glitch with strength power up not properly applying extra damage (health being rounded?)

+++ Assets/Scripts/RayDetector.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class RayDetector : MonoBehaviour {
 
-    public void DetectRayHit(int damage)
+    public void DetectRayHit(float damage)
     {
         SendMessage("ProcessRayHit", damage);
     }
